# Use crypto/subtle.XORBytes