… utils/xor/xor_amd64.go
… utils/xor/xor_amd64.s
… utils/xor/xor_arm64.go
… utils/xor/xor_arm64.s
… utils/xor/xor_ppc64x.go
… utils/xor/xor_ppc64x.s