@stv0g 2024-05-26
Looks good to me :) I think we are ready to bump the minimum Go version to 1.20 by now.

Can you add a nolint-Comment to silence the linter error?