@jech 2024-05-26
> I think we are ready to bump the minimum Go version to 1.20 by now.

Please don't, Debian stable still uses 1.19.

> Can you add a nolint-Comment to silence the linter error?

Done.