@jech 2024-05-26
Just fixed a bug in the xor_old code, which wasn't optimising the AMD64 case correctly.  About to push.